% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_anon_data.R
\name{add_cols}
\alias{add_cols}
\title{Add empty columns to data}
\usage{
add_cols(data, n_to_add, after)
}
\arguments{
\item{data}{Data to add columns to}

\item{n_to_add}{Number of columns to add}

\item{after}{Name of columns to add columns after}
}
\description{
Add empty columns to data
}
