% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cp_number.R
\name{add_cp_number}
\alias{add_cp_number}
\title{Add CP Number}
\usage{
add_cp_number(
  data,
  reg_data,
  age_gender = FALSE,
  age_wave = NULL,
  age_panel = NULL,
  remove_email = FALSE
)
}
\arguments{
\item{data}{Data frame including \code{email}.}

\item{reg_data}{Data frame including \code{email, cp_number}.}

\item{age_gender}{If \code{TRUE}, will also add age group and gender to
\code{data}.}

\item{age_wave}{Wave number of survey to calculate age at.}

\item{age_panel}{Panel of survey to calculate age at.}

\item{remove_email}{If \code{TRUE}, \code{email} will be removed from
returned data.}
}
\value{
\code{data} with \code{cp_number} (and \code{age_group} and
\code{gender} when \code{age_gender = TRUE}) added.
}
\description{
Add CP Number to \code{data} by matching with email address.
}
