% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_group.R
\name{age_group}
\alias{age_group}
\title{Create age groups}
\usage{
age_group(age)
}
\arguments{
\item{age}{A vector of numeric values.}
}
\value{
A character vector, where each element is the age group
for the corresponding element in \code{age}.
}
\description{
\code{age_group()} takes a numeric vector and assigns each
age to the appropriate age group.
}
\examples{
age <- c(27, 30, 54, 89)
age_group(age)

}
