% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymise-data.R
\name{anon_reg_data}
\alias{anon_reg_data}
\title{Anonymise registration data}
\usage{
anon_reg_data(data)
}
\arguments{
\item{data}{Data frame of registration data}
}
\value{
Data frame of registration data with names/nicknames for household
members anonymised.
}
\description{
Anonymise registration data
}
