% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_response_data.R
\name{anon_response_data}
\alias{anon_response_data}
\title{Anonymise response data}
\usage{
anon_response_data(data)
}
\arguments{
\item{data}{Data frame of survey responses}
}
\value{
Data frame of survey responses with names/nicknames for new household
members, existing household members and contacts anonymised.
}
\description{
Anonymise response data
}
