% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymise-data.R
\name{anonymise_data}
\alias{anonymise_data}
\title{Anonymise response data}
\usage{
anonymise_data(data, dataset_to_anon = c("resp", "reg"))
}
\arguments{
\item{data}{Data frame containing data to anonymise}

\item{dataset_to_anon}{Determines which variables should be anonymised.
* "reg": existing household member names/nicknames
* "resp": existing household member names/nicknames, new household member
  names/nicknames, contacts names/nicknames}
}
\value{
Data frame with names/nicknames anonymised.
}
\description{
Anonymise response data
}
