% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_anon_data.R
\name{assign_names}
\alias{assign_names}
\title{Assign names}
\usage{
assign_names(data, names)
}
\arguments{
\item{data}{Data frame to assign names to}

\item{names}{Vector of names to assign}
}
\description{
Assign names
}
