% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_data.R
\name{backup_data}
\alias{backup_data}
\title{Backup data}
\usage{
backup_data(zip_file, file_to_backup)
}
\arguments{
\item{zip_file}{Filepath to zip file; must end with .zip. If this file does
not already exist, it will be created.}

\item{file_to_backup}{Filepath to file to backup}
}
\description{
The function \code{backup_data} adds \code{file_to_backup} to
\code{zip_file}. If \code{zip_file} doesn't exist, it is created and
\code{file_to_backup} is added. If \code{zip_file} does exist,
\code{file_to_backup} is appended to existing zip file.
}
