% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_filepath.R
\name{build_filepath}
\alias{build_filepath}
\title{Build filepath}
\usage{
build_filepath(
  filename,
  survey,
  fileext = ".rds",
  data_folder = here::here("survey-data"),
  create_folder = TRUE
)
}
\arguments{
\item{filename}{Character vector containing file names.}

\item{survey}{Character string; one or two digits to denote survey wave
followed by an optional upper case A or B to denote panel; e.g. 2A, 30B, 50.
This is used to name the sub-directory of \code{data_folder} and to prefix
the file names.}

\item{fileext}{Character string; file extention to append to file name.
Default value is \code{.rds}. If supplied value doesn't begin with \code{.},
then this will be added.}

\item{data_folder}{Character string; full filepath of data folder. This
folder must already exist.}

\item{create_folder}{Logical; Default value of \code{TRUE} creates
sub-directory of \code{data_folder} named using \code{survey} (if it doesn't
already exist). If \code{FALSE}, the sub-folder is not created.}
}
\value{
A character vector of filepaths for the \code{filename}(s) provided.
The filepath begins at \code{data_folder}, followed by a sub-directory called
\code{survey}. Files are given a file extention determined by \code{fileext}.

Note: The folder \code{data_folder/survey} will only be created if
\code{create_folder = TRUE}. The function does not check whether or not the
full filepaths returned exist.

If \code{survey}, \code{fileext}, \code{data_folder} or \code{create_folder}
are more than length 1, then only the first value will be used.
}
\description{
\code{build_filepath} returns a character vector of filepaths.
}
\examples{
build_filepath(filename = c("file1", "file2"),
               survey = "50",
               data_folder = tempdir(),
               create_folder = FALSE)
}
