% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_range}
\alias{date_range}
\title{Date range of survey}
\usage{
date_range(wave, panel = NA_character_, suffix = TRUE)
}
\arguments{
\item{wave}{A numeric value denoting the survey wave number.}

\item{panel}{If wave is 43 or earlier, a character value denoting the survey
panel. Valid options are 'A' and 'B'.}

\item{suffix}{Default value is TRUE. If TRUE, the days will be formatted
with suffix; e.g. 1st, 2nd.}
}
\value{
The date range that the survey was open.
}
\description{
\code{date_range()} takes the wave number and panel
and returns the date range that the survey was open.
}
\examples{
# Date range of survey 24A
date_range(24, "A")

# Date range of survey wave 46 (without suffixes)
date_range(46, suffix = FALSE)

}
