% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files.R
\name{delete_files}
\alias{delete_files}
\title{Delete non-anonymised data files}
\usage{
delete_files(survey)
}
\arguments{
\item{survey}{Character vector of surveys; e.g. 34A, 45. For waves 43 and
earlier, a panel letter (A or B) must be included.}
}
\description{
To comply with data retention rules, non-anonymised data must
be regularly deleted. \code{delete_files} will delete any non-anonymised
data files in data folder for given \code{survey}.
}
