% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_opt_outs}
\alias{dummy_opt_outs}
\title{Dummy opt-out data}
\format{
A data frame with 10 rows and 3 variables including:
\itemize{
  \item Age group
  \item Gender
  \item Number of opt outs
}
}
\usage{
dummy_opt_outs
}
\description{
A dataset containing dummy opt out data for
participants of the Scottish Contact Survey (SCS).
}
\keyword{datasets}
