% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_reg}
\alias{dummy_reg}
\title{Dummy registration data}
\format{
A data frame with 100 rows and 73 variables:
\itemize{
  \item \strong{CP Number}; unique participant identifier.
  \item \strong{Status} of survey participation; 'active', 'reserve'
        or 'opt-out'.
  \item \strong{Panel} of survey; A, B or Z.
  \item \strong{Personal data} including; email, date of birth, gender,
        postcode, local authority, ethnicity, employment status, studying
        status, occupation, occupation of household highest earner, total
        household income, number of vaccine doses received and date this
        information was last updated.
  \item \strong{Household member data} including; name, age, gender,
        occupation and student status for up to ten household members.
}
}
\usage{
dummy_reg
}
\description{
A dataset containing dummy registration data for 100
participants of the Scottish Contact Survey (SCS).
}
\keyword{datasets}
