% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_resp}
\alias{dummy_resp}
\title{Dummy response data}
\format{
A data frame with 50 rows and 45 variables:
\itemize{
  \item \strong{Email} address.
  \item \strong{Household member} name/nickname for up to 10 people.
  \item \strong{New household member} name/nickname for up to 4 people.
  \item \strong{Contact} name/nickname for up to 30 people.
}
}
\usage{
dummy_resp
}
\description{
A dataset containing dummy response data for 50
participants of the Scottish Contact Survey (SCS). \cr

\strong{Note:} \code{dummy_resp} does not contain all variables expected in
the actual response data, but just those required for unit tests. For full
list of variables in actual data, see \code{scs::resp_names}.
}
\keyword{datasets}
