% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_postcode.R
\name{format_postcode}
\alias{format_postcode}
\title{Format a postcode}
\usage{
format_postcode(x)
}
\arguments{
\item{x}{A character string or vector of character strings. Input values
which adhere to the standard UK postcode format may be upper or lower case
and will be formatted regardless of existing spacing. Any input values which
do not adhere to the standard UK postcode format will generate an NA and a
warning message - see \strong{Value} section for more information.}
}
\value{
\code{postcode} returns a character string of length 7. 5 character
postcodes have two spaces after the 2nd character; 6 character postcodes
have 1 space after the 3rd character; and 7 character postcodes have
no spaces.
}
\description{
\code{format_postcode} takes a character string or vector of
character strings. It extracts the input values which adhere to the standard
UK postcode format (with or without spaces), assigns the appropriate amount
of spacing to them and ensures all letters are capitalised.

This function is based on the function of the same name from
Public Health Scotland's
\href{https://github.com/Public-Health-Scotland/phsmethods}{\code{phsmethods}}
package.
}
\details{
The standard UK postcode format (without spaces) is:

\itemize{
\item 1 or 2 letters, followed by
\item 1 number, followed by
\item 1 optional letter or number, followed by
\item 1 number, followed by
\item 2 letters
}
}
