% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cp_number.R
\name{generate_cp_number}
\alias{generate_cp_number}
\title{Generate new CP numbers}
\usage{
generate_cp_number(active_cp, panel, n)
}
\arguments{
\item{active_cp}{A character vector of existing CP numbers.}

\item{panel}{A character indicating which panel the numbers are for.}

\item{n}{An integer indicating how many new CP numbers are required.}
}
\value{
A character vector, length \code{n}, of new CP numbers for
\code{panel}.
}
\description{
\code{generate_cp_number()} creates a vector of new CP numbers
for the required panel.
}
\examples{
x <- c("CPA0001", "CPA0002", "CPB0001")
generate_cp_number(active_cp = x, panel = "A", n = 2)

}
