% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_changes.R
\name{household_changes}
\alias{household_changes}
\title{Household Changes}
\usage{
household_changes(wave, panel, change_type = c("add", "remove"))
}
\arguments{
\item{wave}{Wave of survey data to delete}

\item{panel}{Panel of survey data to delete}
}
\value{
Data frame of household changes from responses for given
\code{wave} and \code{panel}.
}
\description{
Get updates to household members from
anonymised response data for given \code{wave} and \code{panel} of
Scottish Contact Survey (SCS).
}
\examples{
household_changes(36, "A")

}
