% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_opt_outs.R
\name{recode_opt_outs}
\alias{recode_opt_outs}
\title{Recode opt-outs in registration data}
\usage{
recode_opt_outs(reg_data, opt_out_data)
}
\arguments{
\item{reg_data}{A dataframe of registraion data.}

\item{opt_out_data}{A dataframe of cp_numbers opted out.}
}
\value{
Registration data with opt outs recoded.
}
\description{
\code{recode_opt_outs()} takes the registration data, recodes
status for opt outs and removes all registraion data other than
\code{cp_number}, \code{status} and \code{panel}.
}
