% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat-anon-data.R
\name{reformat_anon_reg}
\alias{reformat_anon_reg}
\title{Reformat anonymised registration data}
\usage{
reformat_anon_reg(anon_reg_data, names, wave, panel = NULL)
}
\arguments{
\item{anon_reg_data}{Data frame of anonymised registration data.}

\item{names}{Column names for reformatted data.}

\item{wave}{Survey wave.}

\item{panel}{Survey panel.}
}
\value{
Reformatted data frame of anonymised registration data.
}
\description{
Reformat anonymised registration data
}
