% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_anon_data.R
\name{reformat_anon_resp}
\alias{reformat_anon_resp}
\title{Reformat anonymised response data}
\usage{
reformat_anon_resp(anon_resp_data, names)
}
\arguments{
\item{anon_resp_data}{Data frame of anonymised response data.}

\item{names}{Column names for reformatted data.}
}
\value{
Reformatted data frame of anonymised response data.
}
\description{
Reformat anonymised response data
}
