% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-outs.R
\name{remove_opt_outs}
\alias{remove_opt_outs}
\title{Remove opt-outs from active panel in registration data}
\usage{
remove_opt_outs(reg_data, cp)
}
\arguments{
\item{reg_data}{A dataframe of registration data.}

\item{cp}{A character vector of cp_numbers opted out.}
}
\value{
Registration data with opt outs recoded.
}
\description{
\code{remove_opt_outs()} takes the registration data, recodes
status for opt outs and removes all registraion data other than
\code{cp_number}, \code{status} and \code{panel}.
}
