% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_opt_outs.R
\name{replace_opt_outs}
\alias{replace_opt_outs}
\title{Replace opt outs}
\usage{
replace_opt_outs(reserve_data, opt_out_data)
}
\arguments{
\item{reserve_data}{A data frame containing an age and gender breakdown of
the reserve list.}

\item{opt_out_data}{A data frame containing an age and gender breakdown of
opt outs.}
}
\value{
A character vector of email addresses to replace opt outs.
}
\description{
\code{replace_opt_outs} finds replacements for opt outs from
the reserve list matched by age group and gender
(or age group only if no complete matches).
}
