% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resp_names}
\alias{resp_names}
\title{Response data variable names and types}
\format{
A data frame with 960 rows and 2 variables: \code{names} and
\code{type}.
}
\usage{
resp_names
}
\description{
A dataset containing variable names and types for Scottish
Contact Survey (SCS) response data.
}
\keyword{datasets}
