% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_date.R
\name{start_date}
\alias{start_date}
\title{Calculate first active day of CoMix survey.}
\usage{
start_date(wave, cohort = c("A", "B"))
}
\arguments{
\item{wave}{The number of survey wave in numeric format.}

\item{cohort}{The survey cohort; either A or B.}
}
\value{
Returns a date corresponding to the day the survey opened.
}
\description{
Calculate first active day of CoMix survey.
}
\examples{
# First date of Comix survey 24A
start_date(24, "A")
}
