% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_invites.R
\name{survey_invites}
\alias{survey_invites}
\title{Extract data for survey invites}
\usage{
survey_invites(reg_data, survey_panel)
}
\arguments{
\item{reg_data}{Data frame of registration data}

\item{survey_panel}{Survey panel}
}
\value{
Data frame of participants and registration data required for
survey invites.
}
\description{
Extract data for survey invites
}
