% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_household_members.R
\name{update_household_members}
\alias{update_household_members}
\title{Update household members}
\usage{
update_household_members(reg_data, to_remove = NULL, to_add = NULL)
}
\arguments{
\item{reg_data}{Data frame of registration data}

\item{to_remove}{Data frame of household members that have left household}

\item{to_add}{Data frame of new household members}
}
\value{
A numeric value corresponding to a persons age as at the
beginning of the survey.
}
\description{
\code{update_household_members} takes registration data and
updates household members.
}
