% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_changes.R
\name{vaccine_changes}
\alias{vaccine_changes}
\title{Vaccine Status Changes}
\usage{
vaccine_changes(reg_data, raw_vaccine_data)
}
\arguments{
\item{reg_data}{Registration data}

\item{raw_vaccine_data}{Vaccine data from response data}
}
\value{
Registration data with vaccine status data updated.
}
\description{
Update registration data with latest vaccine status information.
}
