% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Calculate age at time of survey}
\usage{
age(date_of_birth, wave, panel = c("A", "B"))
}
\arguments{
\item{date_of_birth}{Date of birth in date format}

\item{wave}{Wave number}

\item{panel}{Panel number}
}
\value{
A numeric value corresponding to a persons age as at the
beginning of the survey.
}
\description{
\code{age} returns a persons age as at the beginning of a
survey given their date of birth, survey wave number and panel.
}
\examples{
# Age of someone with DOB 25/12/1980 during survey 25A
age(as.Date("1980-12-25"), 25, "A")

}
