% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files.R
\name{delete_files}
\alias{delete_files}
\title{Delete non-anonymised data files}
\usage{
delete_files(wave, panel)
}
\arguments{
\item{wave}{Wave of survey data to delete}

\item{panel}{Panel of survey data to delete}
}
\description{
To comply with data retention rules, non-anonymised data must
be regurlarly deleted. \code{delete_files} will delete any non-anonymised
data files in data folder for given \code{wave} and \code{panel}.
}
