% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_resp}
\alias{dummy_resp}
\title{Dummy response data}
\format{
A data frame with 50 rows and 48 variables including (but not
        limited to):
\itemize{
  \item \strong{CP Number}; unique participant identifier.
  \item \strong{Email} address.
  \item \strong{Household member} name/nickname for up to 10 people.
  \item \strong{Household member changes}; flag to indicate where household
        member has left the household.
  \item \strong{New household member} name/nickname for up to 4 people.
  \item \strong{Contact} name/nickname for up to 30 people.
  \item \strong{Vaccine} dose information.
}
}
\usage{
dummy_resp
}
\description{
A dataset containing dummy response data for 50
participants of the Scottish Contact Survey (SCS).
}
\keyword{datasets}
