% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_changes.R
\name{household_changes}
\alias{household_changes}
\title{Household Changes}
\usage{
household_changes(resp_data, change_type = c("add", "remove"))
}
\arguments{
\item{resp_data}{Data frame of non-anonymised response data.}

\item{change_type}{Type of changes to be extracted; either "add" or "remove.}
}
\value{
Data frame of household changes.
}
\description{
Get updates to household members from
non-anonymised response data for Scottish Contact Survey (SCS).
}
