% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_data_updates.R
\name{reg_data_updates}
\alias{reg_data_updates}
\title{Updates to registration data}
\usage{
reg_data_updates(wave, panel)
}
\arguments{
\item{wave}{Wave of survey data to delete}

\item{panel}{Panel of survey data to delete}
}
\value{
Data frame of updated registration data from responses for given
\code{wave} and \code{panel}.
}
\description{
Get updates to registration data including area of residence,
employment/education status, occupation and household income from
anonymised response data for given \code{wave} and \code{panel} of
Scottish Contact Survey (SCS).
}
