% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-outs.R
\name{replace_opt_outs}
\alias{replace_opt_outs}
\title{Replace opt outs}
\usage{
replace_opt_outs(reg_data, opt_out_data, cur_wave, cur_panel)
}
\arguments{
\item{reg_data}{Data frame of registration data.}

\item{opt_out_data}{Data frame containing an age and gender breakdown of
opt outs.}

\item{cur_wave}{Current wave of the survey}

\item{cur_panel}{Current panel of the survey; panel to add replacements to.}
}
\value{
Data frame of registration data with opt outs replaced.
}
\description{
\code{replace_opt_outs} replaces opt outs from
the reserve list matched by age group and gender
(or age group only if no complete matches).
}
