% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_date.R
\name{start_date}
\alias{start_date}
\title{Date of the first day CoMix survey was open}
\usage{
start_date(wave, panel)
}
\arguments{
\item{wave}{A numeric value denoting the survey wave number.}

\item{panel}{A character value denoting the survey panel. Valid
options are 'A' and 'B'.}
}
\value{
The date of the first day the survey was open.
}
\description{
\code{start_date()} takes the wave number and panel of
a CoMix survey and returns the date of the first day the survey was open.
}
\examples{
# First date of Comix survey 24A
start_date(24, "A")

}
