% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_changes.R
\name{vaccine_changes}
\alias{vaccine_changes}
\title{Vaccine Status Changes}
\usage{
vaccine_changes(wave, panel)
}
\arguments{
\item{wave}{Wave of survey data to delete}

\item{panel}{Panel of survey data to delete}
}
\value{
Data frame of vaccine status changes from responses for given
\code{wave} and \code{panel}.
}
\description{
Get updates to vaccine status from
anonymised response data for given \code{wave} and \code{panel} of
Scottish Contact Survey (SCS).
}
\examples{
vaccine_changes(36, "A")

}
